★引っ越し用/requirements.txt
strands-agents
bedrock-agentcore